requirements.txt
colorama==0.4.1
comet-ml==2.0.18
cupy==6.4.0
Cython==0.29.13
imutils==0.5.3
lmdb==0.98
matplotlib==3.1.1
numpy==1.17.4
pandas==0.25.0
Pillow==6.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.2
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0
tqdm==4.36.1
